refactor: product-info-list config로 이동
