+++ src/features/payment/config/product-info-list.tsx
